add ExtendedCityObjectProvider for selecting and providing cityobjects to View
